\name{localweightedreg-package}
\alias{localweightedreg-package}
\alias{localweightedreg}
\docType{package}
\title{
\packageTitle{localweightedreg}
}
\description{
\packageDescription{localweightedreg}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{localweightedreg}
\packageIndices{localweightedreg}
~~ An overview of how to use the package, including the ~~
~~ most important functions ~~
}
\author{
\packageAuthor{localweightedreg}

Maintainer: \packageMaintainer{localweightedreg}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, ~~
~~ from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
